requirements.txt
crewai
langchain_google_genai
python-dotenv
crewai_tools
litellm
google-generativeai